+++ fabric/src/main/java/com/traverse/taverntokens/registry/FabricModItems.java
@@ -4,8 +4,6 @@ import com.traverse.taverntokens.TavernTokens;
 
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.minecraft.core.Registry;
-import net.minecraft.core.RegistryAccess;
-import net.minecraft.core.RegistryAccess.RegistryEntry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;